requirements.txt
aiogram==3.*
APScheduler==3.*
python-dotenv>=1.0
